Allow promtimer to be started inside unzip'd bundle
There are cases where a user might have a collection of unrelated and
unzip'd cbcollect bundles. Prior to this change it was not possible to
cd into a specific unzip'd directory and run promtimer. This is now
handled by looking for the stats_snapshot directory in the cwd and if
present start promtimer.