Only extract stats_snapshot files from zipped cbcollect ...
... if the stats_snapshot directory is not present.

The old behavior used to extract Prometheus block files if they were
missing. At times Prometheus would consolidate the blocks in the
snapshot causing a subsequent run of Promtimer to extract the original
blocks from the cbcollect zip - which ultimately caused Prometheus to
fail to start with overlapping blocks and Promtimer to crash.