Add ability to get user logs to StatsSource
This patch adds the capability to get user logs to both the
CBCollect and ServerNode stats sources. In the case of CBCollect
the diag log is parsed; in the case of ServerNode the /logs REST
endpoint is called directly. The information that is obtainable
in both cases isn't exactly the same, however, the timestamp
and the text description of the event is the same and so the
interface is useful enough for our purposes.