Add support to create annotations from user logs
Prior to this patch it was only possible to create annotations
from an events.log file that had to be generated using a tool
available in a separate repository. This patch adds support for
creating annotations based on user logs.

User logs are available in the diag.log in cbcollects and can be
directly retrieved from the server in the case that Promtimer is
running against a live cluster. In either case the text description
of the event needs to be examined to determine what kind of event it
is, which is less than ideal but useful enough to add support for it.
Note that the same set of regular expressions is used to identify the
kind of event in both cases.

Additionally this patch adds support for merging new annotations
on top of existing annotation without the creation of duplicates.