Use secureJsonData in data source definition for basic auth info
Despite the fact basicAuthPassword is deprecated, it no longer seems
to work for me on Grafana 8.4 when specifying the password for using
Promtimer against a running cluster. As such moving to use
secureJsonData.basicAuthPassword, which does work.

This highlights another problem of course - we should probably be
using the REST api to create data sources to avoid the data source
definitions sitting on disk.