Extract stats correctly when extra '.' in path
It appears that a recent cbcollect change adds a redundant '.' to the
path that the Prometheus stats snaphot is added to the zip file with.

E.g we now see this in the path in the zip file

cbcollect_info_n_0@127.0.0.1_20220610-011859/./stats_snapshot

whereas previously we saw this:

cbcollect_info_n_0@127.0.0.1_20220610-011859/stats_snapshot

It probably would be preferable to not have this additional, unnecessary
current directory in the path. But it should be the case that Promtimer
should handle it regardless.