Add stats source for metrics from backup archives
This change adds a new type of stats source called BackupStatsFiles that
generates Prometheus metrics from backup archives via the cbmstatparser
utility that is included in the backup repository.

This use of Promtimer can be invoked with the --backup-archive-path
argument and starts Promtimer with a different out-of-the-box dashboard
built off these backup archive metrics.