Containerise Promtimer
The current latest Grafana version available in homebrew (9.4.7) seems
to be incompatible in some way.

At least locally, I could not get it to work properly and kept seeing
"Invalid UID" warnings, 400 HTTP status codes in logs and other
"weird" behaviour which I couldn't debug.

AFAICT homebrew doesn't do precise versioning, so I cannot
`brew install grafana@8` to install an earlier version to fix this.

Instead, I've resorted to running in Docker with this image. To use,
simply `docker build -t promtimer .` in the repo root to create an
image tagged `prometheus` and use the following to mount a dir
containing cbcollects and run the image in a container.

```
docker run --rm -p 13300:13300 -v PATH_TO_CBCOLLECTS:/promtimer/data promtimer
```

This will start a new container with the specified directory mounted in
it and automatically destroy the container on exit. Grafana home will
not be persisted.

If you want your grafana home to be persisted, you can drop the `--rm`
and reuse the container.