Use a hash of the data source name for the data source UID
The UID is constrained to have a max length of 40 characters.
Previously we were just using the data source name for the UID, but this
is obviously problematic in the general case. A simple solution that
retains a deterministic value for the UID is to use a SHA1 hash for it,
the hex representation of which has the additional convenient property
that it is precisely 40 characters.