Use user provided URL when cluster node doesn't have hostname
If we're connecting to a live cluster node that is configured against
the loopback address (i.e. 127.0.0.1), it means the
pools/defaul/nodeServices REST response will not contain a hostname
for this node. To this point, we've been assuming that Promtimer is
running on the same host as the target node - but generally it is
better to use the host that the user provided us to access the node
instead. This patch makes that change.