Connect to Grafana before polling processes and opening browser
There are times when the browser opens and Grafana fails to complete its
startup sequence and the user is left with a blank UI. This is a little
bit annoying usability-wise, so this patch introduces a wait to connect
to the Grafana REST API before opening the browser and polling to see
if the processes are still up. This increases the start-up latency
slightly but overall it seems like the right trade off. Plus the extra
latency makes it more likely that the browser won't open if there's a
different instance of Promtimer running bound to the same ports.