Address issues in PR 41 and get it closer to being mergeable.
* cbbakupmgr-stats.json
The first issue is that the backupmgr dashboard doesn't use the
promtimer dashboard meta model (as described in dashboards/README.md).
One of the points of promtimer is to use the dashboarding templating
idea to keep the dashboard short and easier to undertand and to make
them easy when working with multiple datasources. This may not be an
issue for the backup dashboards for now, but it is certainly the case
that the standard dashboarding model can be used for the backupmgr
so adding new dashboard creation routines just adds unnecessary
complexity.

As such, this patch reworks the cbbackupmgr-stats.json file to be in
standard form. To make it simpler and quicker, I just converted the
first 5 panels. Hopefully this is enough so that it's clear how the
rest of the panels should be converted.

* cbstats.py

Added time_interval - as now it's the case that different stats sources
can use different intervals - and use this method / property in
make_data_sources.

* promtimer.py
Remove the dashboard UID as it's not needed. Instead just refer to it
by its regular name when opening the browser.

Add notion of dashboard_name_predicate so that the main routine can
get the dashboards that are needed created and avoid the others. This is
more general and better than passing a mode everywhere.

Change the bucket template parameter to default to an empty array if
None. A change in the main routine meant that None was getting passed
down in regular situations when promtimer is running against a set of
cbcollects.

Remove make_cbbackupmgr_dashboard as it's not needed.