Refactor stat file timestamp parsing to use datetime
The backup archive mode needs to be able to parse the timestamp from the
last line of the last stat file into a datetime object in order to be
able to set an appropriate time range in the Grafana dashboard.

This commit refactors the previous dateutil implementation into calling
a newly created method that uses datetime instead in order to avoid
adding a new dependency to promtimer.