[KV] Remove misleading kv_memory_used_bytes{for=hashtable}
Pre-trinity, there is a metric called kv_memory_used_bytes with a label
for=hashtable. The name is misleading, it is not the memory used by the
HT. For that reason, we've renamed the metric to kv_ep_kv_size_bytes.

Most of the time, the metric that is wanted is
kv_ep_ht_item_memory_bytes, which is already in the panel, with the same
legend (except it has a capital 'H' in hashtable).

Avoid plotting for=hashtable, to avoid name conflict.