Make dashboards use the same timezone as the logs by default
The dashboards are set to use the browsers timezone by default. This is
perfect for attaching to a live test cluster running locally. For
cbcollects however, we almost always need to remember to change the
timezone.

My personal preference is to always use the timezone that is in the
logs. I've added some extra parsing code which determines the timezone
on the cbcollect host using:
 - /etc/timezone output from couchbase.log
 - cbcollect_info.log timestamps as fallback

In the fallback case, we can only observe an offset from UTC, not a
timezone, which is not enough to determine the timezone that is applied
on the cbcollect host. However, for round hour offsets from UTC, we can
apply an 'Etc' timezone which gives us the same offset with no DST.
In which case an offset of +01:00 maps to Etc/GMT-1.

There are no Etc IANA timezones for non-round offsets, so in that case,
we will fallback to UTC.

The environment variable GRAFANA_TZ can be used to override this an set
a desired value.

Co-Authored-By: Ben Huddleston <ben.huddleston@couchbase.com>