[KV] Make Ephemeral Adjusted Memory panel more accurate
These are metrics we don't expose, but are critical to Ephemeral
auto-delete.

The underlying calculation used is:

const auto estimatedActiveMemory =
        int64_t(stats.getEstimatedTotalMemoryUsed()) -
        stats.inactiveHTMemory - stats.inactiveCheckpointOverhead;

Where inactive means replica + dead vBuckets.

The previous calculation for Pageable Mem Current used an approximation
for the inactiveHTMemory based on: totalHTMemory * inactive VB ratio.

But we have the inactiveHTMemory in stats, as the sum of the per-vBucket
state metrics.

By using that, we get an accurate calculation.

I've also updated the LWM and HWM to use sum without (state) instead of
addition for the two states.