Provide `data-source-uid` to dashboards & use in cbbackupmgr
After upgrading to Grafana 12 the cbbackupmgr dashboard stopped working,
showing "no data". I found that it could be fixed by editing a
visualisation, changing the datasource to something other than
"cbmstatparser-prometheus-tsdb" and then back again.

After some digging it appears that the UI component is getting malformed
data. For the normal cluster dashboards, all of which work, the UI
component gets the source as `{type: "prometheus", uid: "123.."}`. On
the cbbackupmgr ones it gets `"cbmstatparser-prometheus-tsdb"`.

There's no real difference between the dashboards for clusters and the
one for cbbackupmgr that would explain this - both specify the name of
the source, it just seems Grafana is converting one of them to the
object and the other not. Unfortunately my React/TS skills aren't up to
scratch to find where this happens.

To solve this I found that specifying the object in the dashboard itself
fixes it. This commit passes the datasource uid to the dashboard
templates and updates the cbbackupmgr dashboard to use it.