Add support for multi-valued template parameters.
And use this support with a new parameter with name 'data-source' that
supports 'name' and 'uid' attributes.

Prior to this change, parameters were single-valued. E.g. 'bucket'
and 'data-source-name'. However, with the dashboard schema changes in
version 12, it's now the case that referring to datasources by name
only is less reliable. In particular, it now doesn't work for the
cbbackupmgr-stats dashboard for reasons we don't fully understand and
we were required to modify the way data-sources are referred to in that
dashboard as follows:

Before:
            "datasource": "{data-source-name}",

After"
            "datasource": {
              "type": "prometheus",
              "uid": "{data-source-uid}"
            }

This solution works OK for the cbbackupmgr dashboard as it only ever
works with a single datasource. For cases where there are multiple
datasources, we want to be able to refer to both the dashboard UID
_and_ the name. E.g. in cluster-overview we would want to write:

          "datasource": {
              "type": "prometheus",
              "uid": "{data-source-uid}"
          }
          "expr": "sys_cpu_utilization_rate",
          "legendFormat": "{data-source-name} sys_cpu_utilization_rate",

However, if data-source-uid and data-source-name (in the legendFormat)
are independent parameters, the parameter replacement logic will create
their cartesian product and instead of getting N traces on this panel,
we'll get N-squared.

The solution is to support multi-attribute parameters. In this case, we
introduce a new parameter: 'data-source' that has two atributes 'name'
and 'uid' and we write the above as follows:

          "datasource": {
              "type": "prometheus",
              "uid": "{data-source:uid}"
          }
          "expr": "sys_cpu_utilization_rate",
          "legendFormat": "{data-source:name} sys_cpu_utilization_rate",

Where ':' means dereference the given attribute associated with the
parameter value.