Refactor promtimer main function
The entry point to the promtimer script was never great but had degraded
over the years to become an almost incomprehensible collection of
argument parsing, validation and running of the script. It was high time
to clean it up some.

Changes:
* new entry-point function parse_args_validate_and_run that parses and
  does basic validation and invokes a new function: run_promtimer
* run_promtimer does essentially 3 things:
  1) call make_stats_sources
  2) call prepare_grafana
  3) call start_processes_and monitor

This required some small other refactorings. In particular:
a) get_timezone was elevated to be a method in StatsSource and
   the static method get_one_timezone now works with StatsSources
   rather than CBCollects which makes the code a bit simpler
b) handle_backup_archive_mode has been changed to not also return
   a refresh interval (as all it was doing was returning an empty string
   literal) and this is dealt with separately

This code can certainly be further simplified. However, this is a
reasonable first step and good enough for now I think.

Tested all forms of running promtimer before submitting, that is,
running against cbcollects, live clusters and backup archives.

Things can certainly be s