Delete __init__.py
Promtimer doesn't have a module structure. It's just a script and a
bunch of sibling files with reusable functions and classes. At some
point we could consider migrating to some proper module structure,
but right now it seems unnessary.

In the current structure, the __init__.py file is wrong and confuses
some tooling (in particular mypy).