Add grafana-install-path as a command-line argument
Idea of the argument is to allow users run Promtimer under a version of
Grafana which is not their installed Grafana. This is convenient as with
Grafana 12 we've started to see some changes in behavior with the
Promtimer dashboards as a result of the schema changes that are part of
12, so now testing on different versions of Grafana is becoming a more
frequent task.

If the grafana-install-path is provided it will be used to determine the
location of the Grafana binary ($install/bin/grafana-server) and the
Grafana home path ($install/share/grafana, if not otherwise provided.