Fix bug in bucket templating variable drop-down in Grafana 12
The changes in the dashboard schema for Grafana 12 cause the way we lay
out bucket templating variables to break in Grafana 12. 12 expects the
template variable options to be set in a comma-separated list in the
query attribution; versions < 12 require the different variable values
to be laid out as options in the options list. Both versions tolerate
the other way of doing it so the solution is to do both.