Remove use of data-source-name templating variable
And replace with `data-source:name`.

In this previous commit:
* d05d9a3: Add support for multi-valued template parameters

Support for the new `data-source` template parameter with two attributes
`name` and `uid` was added. This patch completes the work started in that
patch with the removal of `data-source-name` as an indepentent template
parameter, replacing it with the `name` attribute of `data-source`.

The replacement in dashboard JSON files is simple: `data-source-name` is
replaced with `data-source:name`. `data-source-name` is now no longer a
valid template parameter to use in dashboards.