Don't rate kv_vb_expired
The stat is not a true counter.

This stat is the aggregate of all vbuckets of that type, vbuckets
can be deleted meaning the stat will drop, a counter only increases.

Instead use the sum of pager/compactor/access - each of these stats
is a counter and safe to apply the rate function to.